requirements.txt
# Core framework
Flask
Werkzeug
gunicorn

# HTTP & webhooks
requests

# FFmpeg Python binding
ffmpeg-python

# S3-compatible storage
boto3

# GCP storage & Cloud Run
google-auth
google-cloud-storage
google-cloud-run
google-api-core

# Payload validation
jsonschema

# Process monitoring
psutil

# Playwright (for /v1/code/execute/python HTML rendering)
playwright

# Media download
yt-dlp